core: imprement logging inside core (read commit)
previously taigabot's stdout log was implemented by a plugin, but since
plugins are susceptible to global ignores and modifications from other
plugins, this meant that the bot output was not showing what was really
going on in irc. sometimes messages would not show up in the terminal
or in the systemd journal. irc lines sent by the bot weren't shown, only
privmsgs to public channels (i think?).

this change implements a custom wrapper around the tcp queues, so that
everything the bot sends/receives is shown in the terminal. the same
timestamp format is kept, but ">>" is added for sent lines and "<<" is
added for incoming lines.

since this "core logging" thing shows the real irc traffic, messages
which were previously "[#chan] <nick> hello" are now shown as the real
irc command "nick!user@host PRIVMSG #chan :hello". i consider it an
improvement but it looks so much worse. TODO fix.

naturally there is a delay in the network, so maybe the lines don't
match up exactly to what the server sees. good enough i guess.

i've filtered the ascii control characters (\x00 up to \x1F) because
they do funny things to the terminal when they're printed.